fix: Linux profiler build + CP doc/log/test correctness cluster
- docs: fix the CP Dockerfile comment -- the no-send guard was removed 2026-07-21
- test: pin exact metric names in the CP drain-success supportability test
- chore: correct CP agent-frame filter claims; log and document disabled CP path
- fix: restore Linux profiler loadability by not ODR-using a CP constant
- fix: catch every mangled spelling in the profiler undefined-symbol guard

+++ src/Agent/NewRelic/Profiler/ContinuousProfiler/ContinuousProfiler.h
@@ -550,7 +550,7 @@ namespace NewRelic { namespace Profiler { namespace ContinuousProfiler
                 // persistent slots hold, so the extras were dropped rather than growing the capture buffer
                 // under suspend.
                 LogTrace(L"CP: thread capture overflow; dropped ", overflowCount, L" thread(s) beyond the ",
-                    ThreadCountForReservation, L"-slot capture buffer");
+                    static_cast<size_t>(ThreadCountForReservation), L"-slot capture buffer");
             }
 
             EncodeAndPublish(failedSnapshotCount, batchTimestamp, microsSuspended);
